pyproject.toml
[project]
name = "mcpt"
version = "0.2.1"
description = "CLI for discovering and running MCP Tool Shop tools"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "mcp-tool-shop", email = "64996768+mcp-tool-shop@users.noreply.github.com"}
]
keywords = ["mcp", "cli", "tools", "registry", "discovery", "package-manager"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "rich>=13.0",
    "platformdirs>=4.0",
    "httpx>=0.27",
]

[project.urls]
Homepage = "https://github.com/mcp-tool-shop-org/mcpt"
Repository = "https://github.com/mcp-tool-shop-org/mcpt"
Issues = "https://github.com/mcp-tool-shop-org/mcpt/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[project.scripts]
mcpt = "mcpt.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["mcpt"]
